refactor(fresco): split NAPI type modules
